Initial homebank-db crate with basic schema
This should cover the basics and still be very flexible.
I could take a few routes from here: map HomeBank's source code from C/C++ to Rust, or reimplement a bunch of stuff.
Either route should still be consistent with what I have for now.

Naively, everything is a bunch of `Vec`s, but this should ideally be a database of some kind.