Adding some tests for date conversion
There are some off-by-1 or -2 errors with the date conversion.
I think this has to do with the switch between Julian and Gregorian calendars and/or leap days, but I'm not sure.
The date that I see in HomeBank is not the same as the value I parse.
These failing tests help highlight that.