Allow querying transfers without specifying values
Previous implementation of filtering `Transaction`s in a query
didn't work anymore for the new implementation of a `Transfer`.
Added a method to `TransactionType` to check if two types are similar.
Didn't want to implement custom versions of `PartialEq` or `Eq`
since that would be complicated and interfere with other operations
on `Transfer`s.

This allows the user to still select `Transfer`s from the CLI, like
before the `Transfer` redesign.