Storing duplicate data in `SimpleTransaction`
This duplication of the amount and memo does come at a cost, but saves on the overall complexity of introducing lifetimes everywhere in the code base.
Duplicating this data also allows for more consistent access between `Simple` and `Split` transactions.

Updating some methods to produced `Vec` of pointers instead of pointers to `Vec`.