Simplifying `Query` lifetimes
Since we're creating `Transaction`s on the fly, this actually creates new values.
It doesn't just store pointers.
While possibly computationally expensive, we leave these values to the end of the `Query::exec` method to reduce the potential impact.

Because of this, we can remove the lifetime parameters from the `Query::exec` method.
This also requires the addition of `Clone` implementations to various structs.

Also refactoring `GroupError` into its own module for consistency with other types.