Query categories by their full name
For top-level categories, nothing changes.
For sub-categories, this changes a few things.
The full name is displayed as `{parent}:{sub}`, so regexes that previously matched the beginning of a word will no longer work.
But this does make it easier to distinguish between sub-categories with the same name that have different parents.
For example, it was previously impossible to only select the `Hello:There` category and not the `WhoGoes:There` category.
Now you can.