Completed compatibility with `AccountFlags`
Methods and implementations to handle the conversion between types.
Also included some helper functions on the `AccountFlags` type to make the comparisons more explicit in code than `acc->flags & AF_CLOSED`.

I'm not sure if this implementation is the best, because it does feel a bit clunky.
But I'll try it and see how it feels.