Removing `Account::new()` method
I don't actually use this method anywhere in the code base.
It's not necessary, but it is part of the `homebank-db` library.
So theoretically, if someone was using this library as a dependency, this would be a breaking change.

But no one is using this crate, and it's still pre-1.0.0, so I feel comfortable making this change.
I also didn't like how many arguments this method had, and neither did clippy.